requirements.txt
pyspark==3.5.1
jupyter
kafka-python